Monkey-patch database authenticatable
